+++ core-ui/src/header/Header.scss
@@ -17,3 +17,7 @@
     }
   }
 }
+
+.fd-shellbar--l {
+  padding-left: 0 !important;
+}
