+++ core-ui/src/header/Logo/Logo.scss
@@ -1,0 +1,10 @@
+.fd-shellbar__logo {
+  display: flex;
+  gap: 10px;
+  justify-content: center;
+  align-items: center;
+
+  img {
+    height: 100%;
+  }
+}
