+++ core-ui/src/header/SidebarSwitcher/SidebarSwitcher.scss
@@ -1,19 +1,4 @@
-.fd-shellbar--l {
-  padding-left: 0 !important;
-}
-
-.fd-shellbar__logo {
-  display: flex;
-  gap: 10px;
-  justify-content: center;
-  align-items: center;
-
-  img {
-    height: 100%;
-  }
-}
-
-.nav__buton-wrapper {
+.sidebar-switcher-wrapper {
   display: flex;
   justify-content: center;
   align-items: center;
